Work on clipboard implementation
