+++ mfp/gui/clutter/app_window.py
@@ -520,47 +520,6 @@ class ClutterAppWindowImpl (AppWindow, AppWindowImpl):
         if cliptxt:
             clipboard.set_text(cliptxt, -1)
 
-    async def clipboard_cut(self, pointer_pos):
-        if self.selected:
-            await self.clipboard_copy(pointer_pos)
-            await self.delete_selected()
-            return True
-        return False
-
-    async def clipboard_copy(self, pointer_pos):
-        if self.selected:
-            cliptxt = await MFPGUI().mfp.clipboard_copy(
-                pointer_pos,
-                [o.obj_id for o in self.selected if o.obj_id is not None]
-            )
-            self.clipboard_set(cliptxt)
-            return True
-        return False
-
-    async def clipboard_paste(self, pointer_pos=None):
-        cliptxt = self.clipboard_get()
-        if not cliptxt:
-            return False
-
-        newobj = await MFPGUI().mfp.clipboard_paste(
-            cliptxt, self.selected_patch.obj_id,
-            self.selected_layer.scope, None
-        )
-
-        if newobj is not None:
-            await self.unselect_all()
-            for o in newobj:
-                obj = MFPGUI().recall(o)
-                if obj is None:
-                    return True
-                if not isinstance(obj, PatchDisplay):
-                    obj.move_to_layer(self.selected_layer)
-                    if obj not in self.selected:
-                        await self.select(MFPGUI().recall(o))
-            return False
-        else:
-            return False
-
     def screen_to_canvas(self, x, y):
         success, new_x, new_y = self.group.transform_stage_point(x, y)
         if success:
